refactor: Command 분기문 제거를 위한 CommandContoller 생성
