+++ src/main/java/controller/CommandType.java
@@ -1,0 +1,8 @@
+package controller;
+
+public enum CommandType {
+
+    MOVE,
+    PASS,
+    SURRENDER
+}
