+++ src/main/java/controller/GameCommandFactory.java
@@ -1,0 +1,12 @@
+package controller;
+
+public class GameCommandFactory {
+
+    public static GameCommand create(CommandType type) {
+        return switch (type) {
+            case MOVE -> new MoveController();
+            case PASS -> new PassController();
+            case SURRENDER -> new SurrenderController();
+        };
+    }
+}
